Fixed Demo
Got .nvmrc file updated to latest version of the demo to use node 14.2.0.  I also made a similar update to the Dockerfile as well.  I also changed the port in the docker-compose.yaml from using the internal Docker container port of 1234 to 3005.  This will make it easier to access the demo page once the docker container is running.